Make interpolation (X,Y) the same length
When interpolating wind speeds over altitudes, the length of the
altitude and windspeed data are not always the same. This commit tries
to do the best of the situation and discard the most high altitude data.

The fact that they do not match is a sign that there is something wrong
with the data, and the results are probably slightly wrong.